# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Restful API Documentation",
        "description": "Restful API documentation",
        "contact": {
            "email": "qrizan@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Base Restful API Documentation"
        }
    ],
    "paths": {
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get users",
                "operationId": "9ded17315e13756cd593b5840cd39a26",
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Store new user data",
                "operationId": "f6f90d5fbbe24f66eb2b43901e0d3f38",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password user",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Password user",
                                        "type": "string"
                                    },
                                    "roles": {
                                        "description": "Password user",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "test admin",
                                    "email": "admin@email.com",
                                    "password": "Password123$",
                                    "password_confirmation": "Password123$",
                                    "roles": [
                                        "marketing",
                                        "roles",
                                        "reviewer"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get user detail by id",
                "operationId": "c6ac2250cb22823f621026268b66f471",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update user by id",
                "operationId": "22692373e5a35c609175c1e6441b58ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password user",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Password user",
                                        "type": "string"
                                    },
                                    "roles": {
                                        "description": "Password user",
                                        "type": "array",
                                        "items": {
                                            "description": "The survey ID",
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "test admin",
                                    "email": "admin@email.com",
                                    "password": "Password123$",
                                    "password_confirmation": "Password123$",
                                    "roles": [
                                        "marketing",
                                        "roles",
                                        "reviewer"
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remove user by id",
                "operationId": "fdbad6312957544572f3b379b96abad0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}